sgm-75: refactor: move MovieRemote and MovieMapper to movie package and add tmdbId
- Moves `MovieRemote` and `MovieMapper` classes to the `movie` package.
- Adds `tmdbId` property to `MovieRemote`, `MovieMapper`, and `MovieMockDataSourceImpl`.
- Updates references to `MovieRemote` and `MovieMapper` across the codebase.
