+++ data/src/main/java/io/kikiriki/sgmovie/data/repository/movie/remote/MovieRemoteDataSource.kt
@@ -1,6 +1,6 @@
 package io.kikiriki.sgmovie.data.repository.movie.remote
 
-import io.kikiriki.sgmovie.data.model.MovieRemote
+import io.kikiriki.sgmovie.data.model.movie.MovieRemote
 
 interface MovieRemoteDataSource {
     suspend fun get(lang: String, coproductions: Boolean = false) : List<MovieRemote>
